requirements.txt

python-dotenv
PyPDF2
fastapi
uvicorn[standard]
sqlalchemy
pydantic-ai-slim[duckduckgo]
duckduckgo-search>=5.0.0
supabase
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

pydantic[email]
pydantic
pydantic_ai
langchain
langgraph
psycopg2-binary
PyJWT
langchain-google-genai
urllib3
langchain-community
python-jose[cryptography]
praw
mastodon.py

neo4j
supabase
alembic
qdrant-client
tweepy
google-analytics-data

pytrends
newsapi-python
asyncpraw